speed up the sparse features

+++ ModelUpdate.h
@@ -48,7 +48,7 @@ public:
 		for (int idx = 0; idx < _params.size(); idx++){
 			sumNorm += _params[idx]->squareGradNorm();
 		}
-		if (std::isnan(sumNorm) || sumNorm > 1e20){ //too large
+		if (std::isnan(double(sumNorm)) || sumNorm > 1e20){ //too large
 			clearGrad();
 			return;
 		}
